supertest testing

+++ src/models/Product.model.ts
@@ -13,7 +13,7 @@ export class Product extends Model {
     declare name: string
 
     @Column({
-        type:DataType.FLOAT(6,2)
+        type:DataType.FLOAT(6)
     })
     declare price: number
